Added Signup screen, connected redux, added navigation.
Signup screen: Added a screen for signup. Logic not implemented yet.
Routing and navigation: Added reactnavigation to the project, setup the basic navigation between the three screens.
Created a placeholder for the NewUpdate screen.
Setup a redux store and connected it to track the logged status of the user. The logged status is used for the navigation.

Ref: #001