Added automatic update to the date display on top of the update page
Added the functionality to get current date, yesterday and tomorrow and update it in the date's components showing the current date.
Currently this is only a static display and I aim to replace it in the future with a carrousel scroll so the user can potentially edit other days when forgotten.

Ref: Req #016